Move string tests to a standalone test
